fix: Preserve decimal and negative values for letterSpacing in typography parser
Previously, letterSpacing was parsed using asInt(0), causing all decimal or negative values to be rounded or zeroed out.
Switched to asDouble(0.0) and cast to float to correctly support fractional spacing values.

+++ src/com/hridoy/thmlt/ThMLT.java
@@ -526,7 +526,7 @@ public class ThMLT extends AndroidNonvisibleComponent {
         // Defensive checks for missing fields or type mismatches
         int fontSize = typographyValueNode.path("fontSize").asInt(0);
         int lineHeight = typographyValueNode.path("lineHeight").asInt(0);
-        float letterSpacing = typographyValueNode.path("letterSpacing").asInt(0);
+        float letterSpacing = (float) typographyValueNode.path("letterSpacing").asDouble(0.0);
         String linkedFont = typographyValueNode.path("linkedFont").asText();
 
         TypographyTemplate typography = new TypographyTemplate(fontSize, lineHeight, letterSpacing, linkedFont);
